Cast mel spectrogram to the waveform dtype in AudioVAE encode.
waveform_to_mel returns its own dtype, which then reaches
autoencoder.encode unconverted and mismatches the VAE's weights.

+++ comfy/ldm/lightricks/vae/audio_vae.py
@@ -152,7 +152,7 @@ class AudioVAE(torch.nn.Module):
 
         mel_spec = self.preprocessor.waveform_to_mel(
             waveform, waveform_sample_rate, device=waveform.device
-        )
+        ).to(dtype=waveform.dtype)
 
         latents = self.autoencoder.encode(mel_spec)
         posterior = DiagonalGaussianDistribution(latents)
